use kubeClient context for connecting to kube api-server rather creating new clients everytime, Also, rename the filenames
