+++ cli/cmd/metrics/chunk_interval_get.go
