+++ cli/cmd/prometheus/port_forward.go
@@ -31,12 +31,13 @@ func prometheusPortForward(cmd *cobra.Command, args []string) error {
 		return fmt.Errorf("could not port-forward Prometheus: %w", err)
 	}
 
-	serviceName, err := k8s.KubeGetServiceName(root.Namespace, map[string]string{"release": root.HelmReleaseName, "app": "kube-prometheus-stack-prometheus"})
+	k8sClient := k8s.NewClient()
+	serviceName, err := k8sClient.KubeGetServiceName(root.Namespace, map[string]string{"release": root.HelmReleaseName, "app": "kube-prometheus-stack-prometheus"})
 	if err != nil {
 		return fmt.Errorf("could not port-forward Prometheus: %w", err)
 	}
 
-	_, err = k8s.KubePortForwardService(root.Namespace, serviceName, port, common.FORWARD_PORT_PROM)
+	_, err = k8sClient.KubePortForwardService(root.Namespace, serviceName, port, common.FORWARD_PORT_PROM)
 	if err != nil {
 		return fmt.Errorf("could not port-forward Prometheus: %w", err)
 	}
