+++ cli/cmd/timescaledb/port_forward.go
@@ -31,12 +31,13 @@ func timescaledbPortForward(cmd *cobra.Command, args []string) error {
 		return fmt.Errorf("could not port-forward TimescaleDB: %w", err)
 	}
 
-	podName, err := k8s.KubeGetPodName(root.Namespace, map[string]string{"release": root.HelmReleaseName, "role": "master"})
+	k8sClient := k8s.NewClient()
+	podName, err := k8sClient.KubeGetPodName(root.Namespace, map[string]string{"release": root.HelmReleaseName, "role": "master"})
 	if err != nil {
 		return fmt.Errorf("could not port-forward TimescaleDB: %w", err)
 	}
 
-	_, err = k8s.KubePortForwardPod(root.Namespace, podName, port, common.FORWARD_PORT_TSDB)
+	_, err = k8sClient.KubePortForwardPod(root.Namespace, podName, port, common.FORWARD_PORT_TSDB)
 	if err != nil {
 		return fmt.Errorf("could not port-forward TimescaleDB: %w", err)
 	}
