+++ cli/cmd/volume/get.go
@@ -50,8 +50,9 @@ func volumeGet(cmd *cobra.Command, args []string) error {
 		promStorage, tsDBStorage, tsDBWal = true, true, true
 	}
 
+	k8sClient := k8s.NewClient()
 	if tsDBStorage {
-		results, err := k8s.GetPVCSizes(root.Namespace, pvcStorage, common.GetTimescaleDBLabels())
+		results, err := k8sClient.GetPVCSizes(root.Namespace, pvcStorage, common.GetTimescaleDBLabels())
 		if err != nil {
 			return fmt.Errorf("could not get timescaleDB-storage: %w", err)
 		}
@@ -59,7 +60,7 @@ func volumeGet(cmd *cobra.Command, args []string) error {
 	}
 
 	if tsDBWal {
-		results, err := k8s.GetPVCSizes(root.Namespace, pvcWAL, common.GetTimescaleDBLabels())
+		results, err := k8sClient.GetPVCSizes(root.Namespace, pvcWAL, common.GetTimescaleDBLabels())
 		if err != nil {
 			return fmt.Errorf("could not get timescaleDB-wal: %w", err)
 		}
@@ -67,7 +68,7 @@ func volumeGet(cmd *cobra.Command, args []string) error {
 	}
 
 	if promStorage {
-		results, err := k8s.GetPVCSizes(root.Namespace, pvcPrometheus, common.GetPrometheusLabels())
+		results, err := k8sClient.GetPVCSizes(root.Namespace, pvcPrometheus, common.GetPrometheusLabels())
 		if err != nil {
 			return fmt.Errorf("could not get prometheus-storage: %w", err)
 		}
