+++ cli/pkg/timescaledb_secrets/create.go
@@ -25,13 +25,14 @@ type TSDBSecretsInfo struct {
 	EnableS3Backup bool
 	TlsCert        []byte
 	TlsKey         []byte
+	K8sClient      k8s.Client
 }
 
 func (t *TSDBSecretsInfo) CreateTimescaleDBSecrets() error {
 	// Previous helm install used to create namespace if it doesn't exist
 	// but as we are creating secrets prior to deploying tobs. We are verifying
 	// namespace if doesn't create one.
-	err := k8s.CreateNamespaceIfNotExists(t.Namespace)
+	err := t.K8sClient.CreateNamespaceIfNotExists(t.Namespace)
 	if err != nil {
 		return err
 	}
@@ -79,7 +80,7 @@ func randomPassword(n int) ([]byte, error) {
 
 func (t *TSDBSecretsInfo) createTimescaleDBCredentials() error {
 	secretName := t.ReleaseName + "-credentials"
-	exists, err := k8s.CheckSecretExists(secretName, t.Namespace)
+	exists, err := t.K8sClient.CheckSecretExists(secretName, t.Namespace)
 	if err != nil {
 		return err
 	}
@@ -116,7 +117,7 @@ func (t *TSDBSecretsInfo) createTimescaleDBCredentials() error {
 		Type: "Opaque",
 	}
 
-	return k8s.CreateSecret(sec)
+	return t.K8sClient.CreateSecret(sec)
 }
 
 func generateCerts() ([]byte, []byte, error) {
@@ -179,7 +180,7 @@ WARNING: Using a generated self-signed certificate for TLS access to TimescaleDB
 
 `)
 	secretName := t.ReleaseName + "-certificate"
-	exists, err := k8s.CheckSecretExists(secretName, t.Namespace)
+	exists, err := t.K8sClient.CheckSecretExists(secretName, t.Namespace)
 	if err != nil {
 		return err
 	}
@@ -205,7 +206,7 @@ WARNING: Using a generated self-signed certificate for TLS access to TimescaleDB
 		Type: "Opaque",
 	}
 
-	return k8s.CreateSecret(sec)
+	return t.K8sClient.CreateSecret(sec)
 }
 
 func (t *TSDBSecretsInfo) createS3BackupForTimescaleDB() error {
@@ -270,7 +271,7 @@ Google Cloud:
 	}
 
 	fmt.Println()
-	err := createTimescaleDBPgBackRest(t.ReleaseName, t.Namespace, s3)
+	err := t.createTimescaleDBPgBackRest(t.ReleaseName, t.Namespace, s3)
 	if err != nil {
 		return err
 	}
@@ -278,9 +279,9 @@ Google Cloud:
 	return nil
 }
 
-func createTimescaleDBPgBackRest(name, namespace string, s3 s3Details) error {
+func (t *TSDBSecretsInfo) createTimescaleDBPgBackRest(name, namespace string, s3 s3Details) error {
 	secretName := name + "-pgbackrest"
-	exists, err := k8s.CheckSecretExists(secretName, namespace)
+	exists, err := t.K8sClient.CheckSecretExists(secretName, namespace)
 	if err != nil {
 		return err
 	}
@@ -315,7 +316,7 @@ func createTimescaleDBPgBackRest(name, namespace string, s3 s3Details) error {
 		Type: "Opaque",
 	}
 
-	return k8s.CreateSecret(sec)
+	return t.K8sClient.CreateSecret(sec)
 }
 
 func pemEncodeCert(certDerBytes []byte) ([]byte, error) {
