+++ cli/pkg/timescaledb_secrets/delete.go
@@ -6,14 +6,14 @@ import (
 	"github.com/timescale/tobs/cli/pkg/k8s"
 )
 
-func DeleteTimescaleDBSecrets(releaseName, namespace string, backUpEnabled bool) {
+func DeleteTimescaleDBSecrets(k8sClient k8s.Client, releaseName, namespace string, backUpEnabled bool) {
 	fmt.Println("Deleting TimescaleDB secrets...")
 	credentialsSecret := []string{releaseName + "-credentials", releaseName + "-certificate"}
 	if backUpEnabled {
 		credentialsSecret = append(credentialsSecret, releaseName+"-pgbackrest")
 	}
 	for _, s := range credentialsSecret {
-		err := k8s.DeleteSecret(s, namespace)
+		err := k8sClient.DeleteSecret(s, namespace)
 		if err != nil {
 			fmt.Printf("failed to delete %s secret %v\n", s, err)
 		}
