+++ cli/tests/tobs-cli-tests/metrics_test.go
@@ -242,7 +242,8 @@ func TestMetrics(t *testing.T) {
 		t.Skip("Skipping metrics tests")
 	}
 
-	secret, err := k8s.KubeGetSecret(NAMESPACE, RELEASE_NAME+"-credentials")
+	k8sClient := k8s.NewClient()
+	secret, err := k8sClient.KubeGetSecret(NAMESPACE, RELEASE_NAME+"-credentials")
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -254,7 +255,7 @@ func TestMetrics(t *testing.T) {
 	}
 
 	dbLabelsSet := map[string]string{"release": RELEASE_NAME, "role": "master"}
-	podName, err := k8s.KubeGetPodName(NAMESPACE, dbLabelsSet)
+	podName, err := k8sClient.KubeGetPodName(NAMESPACE, dbLabelsSet)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -265,7 +266,7 @@ func TestMetrics(t *testing.T) {
 
 	stdout := os.Stdout
 	os.Stdout = os.NewFile(0, os.DevNull)
-	_, err = k8s.KubePortForwardPod(NAMESPACE, podName, 5433, 5432)
+	_, err = k8sClient.KubePortForwardPod(NAMESPACE, podName, 5433, 5432)
 	if err != nil {
 		t.Fatal(err)
 	}
